requirements.txt
langchain
langchain-core
langchain-google-genai
langgraph
langsmith
pinecone
fastapi
uvicorn
python-dotenv
ollama
